2 datepicker inputs for 1 calendar
Remove fake 3rd input for instantiation and showing of datepicker.
It made a feel that you can click on any of two inputs and show calendar, but really you always clicked on one invisible 3rd input that was placed on top of those two inputs and they weren't clickable really.
Now it is really two clickable inputs and clicking on first or second input shows the same calendar instance.